requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
pytest==7.4.2
pytest-cov==4.1.0
